Update random_walk.tex
In the code of this exercise you used X_k = {1,2,3} and S_k = {1,2,4}. Then if you run the code and look at the plot, after some 10 customers the distribution hardly changes any further. But if you run the code with X_k = {1,3} and S_k = {1,2} after some 30 customers the distribution hardly changes. Since we used X_k = {1,2,3} and S_k = {1,2,4} also in the example in the text, I assumed that this was right type. Or the solution is not correct