Update gordon_newell.tex
First I found a mistake made in line 110 since V_5 = V_3 \cdot 2/3 and not V_5 = V_4 \cdot 2/3. But after calculating the result I started with V_6 = 1 and ended with V_6 = 48/45 > 1. So I followed the steps and found some other errors, so I changed the solution and I added another way (stated between the brackets) to obtain the results.